refactor: keep copy button reusable
